feat(FN-1088): rename file which creates on task node events
